% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-msi-row-numbers.R
\name{mutate_msi_row_numbers}
\alias{mutate_msi_row_numbers}
\title{Add row numbers to the locus table section of a CLC MSI Report}
\usage{
mutate_msi_row_numbers(
  df,
  regex_table = "Percentage: Baseline",
  regex_locus = ".*\\\\(\\\\w{1,2}\\\\)\\\\d{1,3}",
  regex_stability = "Stability"
)
}
\arguments{
\item{df}{The dataframe output from read_msi_locus_section}

\item{regex_table}{The column name of the final column of each locus
length table}

\item{regex_locus}{Regular expression for identifying the locus name, which
is 2 rows above the corresponding length table}

\item{regex_stability}{Regular expression for identifying the cell containing
the stability prediction for a locus}
}
\value{
The input dataframe with added columns showing the rownumber of
each cell
}
\description{
Add row numbers to the locus table section of a CLC MSI Report
}
